I made my own version of regress.sh! It's called test.sh. I removed the redundant function retSize and replaced calls to it with typeToBytes

+++ MJCPA5/test/Passing/Access1.java
@@ -25,11 +25,11 @@ class C{
 
 	public int test2(){
 		int[] AI;	
-		AI = new int[3];
+		AI = new int[(byte)3];
 		
 		Meggy.delay(AI[0]);
 
-		return ia[1];
+		return ia[(byte)1];
 	}
 
 	public Meggy.Color test3(){
